Make Pickers inherit from AppCompatDialogFragment
AppCompatDialogFragment inherits from DialogFragment, so this is not an API breaking change.
However, styling of AppCompatDialogFragments works slightly different, so we should put a disclaimer in the release notes.
The main advantage is that "dialogCornerRadius" is now correctly applied (previously only the API 28 attribute "android:dialogCornerRadius" would work)
Fixes #588
Fixes #504